requirements.txt
APScheduler==3.9.1
beautifulsoup4==4.9.2
certifi==2020.6.20
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cmake==3.27.7
cryptography==41.0.5
decorator==4.4.2
dnspython==2.0.0
idna==2.10
pycparser==2.20
pymongo==3.11.0
python-dotenv==0.20.0
python-telegram-bot==12.8
pytz==2023.3.post1
requests==2.31.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
tzlocal==5.2
urllib3==1.26.5
